pyproject.toml
[tool.poetry]
name = "capricci-backend"
version = "1.0.0"
description = ""
authors = ["Pol Lopez <pol.lopez.cano@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11.9"
django = "^5.0.4"
pre-commit = "^3.7.0"
django-ninja = "^1.1.0"
django-environ = "^0.11.2"
psycopg2-binary = "^2.9.9"
whitenoise = {extras = ["brotli"], version = "^6.6.0"}
uvicorn = "^0.29.0"
gunicorn = "^22.0.0"
pillow = "^10.3.0"
django-ninja-extra = "^0.20.5"
django-storages = "^1.14.2"
boto3 = "^1.34.93"
faker = "^24.14.1"
randimage = "^1.2"
django-cors-headers = "^4.3.1"
django-ninja-jwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
line-length = 120
indent-width = 4
target-version = "py311"

[tool.ruff.lint]
select = [
  "E",
  "F",
  "W",
  "I",
  "N",
  "UP",
  "FBT",
  "B",
  "A",
  "C4",
  "DJ",
  "EM",
  "ICN",
  "LOG",
  "G",
  "PIE",
  "PT",
  "RSE",
  "RET",
  "SIM",
  "ARG",
  "PTH",
  "PL",
  "PERF",
  "RUF",
  ]
ignore = ["RUF012", "G004", "ARG001"]
fixable = ["ALL"]
exclude = ["*/migrations/*"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
exclude = ["*/migrations/*"]
